Lift ListBuckets state
ListBuckets handled state and bucket deletion, which didn't really make sense and ultimately resulted in a state error when trying to delete a bucket when numBuckets == 1.
