+++ src/features/ObjectStorage/Buckets/ListBuckets.tsx
@@ -1,8 +1,5 @@
 import * as React from 'react';
 import { compose } from 'recompose';
-import ActionsPanel from 'src/components/ActionsPanel';
-import Button from 'src/components/Button';
-import ConfirmationDialog from 'src/components/ConfirmationDialog';
 import Paper from 'src/components/core/Paper';
 import {
   createStyles,
@@ -12,21 +9,12 @@ import {
 } from 'src/components/core/styles';
 import TableBody from 'src/components/core/TableBody';
 import TableHead from 'src/components/core/TableHead';
-import Typography from 'src/components/core/Typography';
 import Paginate from 'src/components/Paginate';
 import PaginationFooter from 'src/components/PaginationFooter';
 import Table from 'src/components/Table';
 import TableCell from 'src/components/TableCell';
 import TableRow from 'src/components/TableRow';
 import TableSortCell from 'src/components/TableSortCell';
-import TextField from 'src/components/TextField';
-import bucketRequestsContainer, {
-  BucketsRequests
-} from 'src/containers/bucketRequests.container';
-import useOpenClose from 'src/hooks/useOpenClose';
-import { getErrorStringOrDefault } from 'src/utilities/errorUtils';
-import { sendDeleteBucketEvent } from 'src/utilities/ga';
-import { readableBytes } from 'src/utilities/unitConversions';
 import BucketTableRow from './BucketTableRow';
 
 type ClassNames = 'root' | 'label' | 'confirmationCopy';
@@ -36,9 +24,6 @@ const styles = (theme: Theme) =>
     root: {},
     label: {
       paddingLeft: 65
-    },
-    confirmationCopy: {
-      marginTop: theme.spacing(1)
     }
   });
 
@@ -47,107 +32,20 @@ interface Props {
   orderBy: string;
   order: 'asc' | 'desc';
   handleOrderChange: (orderBy: string, order?: 'asc' | 'desc') => void;
+  handleClickRemove: (bucket: Linode.Bucket) => void;
 }
 
-type CombinedProps = Props & WithStyles<ClassNames> & BucketsRequests;
+type CombinedProps = Props & WithStyles<ClassNames>;
 
 export const ListBuckets: React.StatelessComponent<CombinedProps> = props => {
-  const { data, orderBy, order, handleOrderChange, classes } = props;
-
-  const removeBucketConfirmationDialog = useOpenClose();
-  const [bucketToRemove, setBucketToRemove] = React.useState<
-    Linode.Bucket | undefined
-  >(undefined);
-  const [isLoading, setIsLoading] = React.useState<boolean>(false);
-  const [error, setError] = React.useState<string>('');
-  const [confirmBucketName, setConfirmBucketName] = React.useState<string>('');
-
-  const handleClickRemove = (bucket: Linode.Bucket) => {
-    setBucketToRemove(bucket);
-    setError('');
-    removeBucketConfirmationDialog.open();
-  };
-
-  const removeBucket = () => {
-    const { deleteBucket } = props;
-
-    // This shouldn't happen, but just in case (and to get TS to quit complaining...)
-    if (!bucketToRemove) {
-      return;
-    }
-
-    setError('');
-    setIsLoading(true);
-
-    const { cluster, label } = bucketToRemove;
-    // Passing in `force: 1` as a param to delete ALL items within
-    // the bucket before deleting the bucket itself.
-    deleteBucket({ cluster, label, params: { force: 1 } })
-      .then(() => {
-        removeBucketConfirmationDialog.close();
-        setIsLoading(false);
-
-        // @analytics
-        sendDeleteBucketEvent(cluster);
-      })
-      .catch(e => {
-        setIsLoading(false);
-        const errorText = getErrorStringOrDefault(e, 'Error removing bucket.');
-        setError(errorText);
-      });
-  };
-
-  const actions = () => (
-    <ActionsPanel>
-      <Button
-        buttonType="cancel"
-        onClick={() => {
-          removeBucketConfirmationDialog.close();
-        }}
-        data-qa-cancel
-      >
-        Cancel
-      </Button>
-      <Button
-        buttonType="secondary"
-        destructive
-        onClick={removeBucket}
-        data-qa-submit-rebuild
-        loading={isLoading}
-        disabled={
-          bucketToRemove ? confirmBucketName !== bucketToRemove.label : true
-        }
-      >
-        Delete
-      </Button>
-    </ActionsPanel>
-  );
-
-  const deleteBucketConfirmationMessage = bucketToRemove ? (
-    <React.Fragment>
-      {bucketToRemove.size > 0 ? (
-        <Typography>
-          This bucket contains{' '}
-          <strong>
-            {bucketToRemove.objects}{' '}
-            {bucketToRemove.objects === 1 ? 'object' : 'objects'}
-          </strong>{' '}
-          totalling{' '}
-          <strong>{readableBytes(bucketToRemove.size).formatted}</strong> that
-          will be deleted along with the bucket. Deleting a bucket is permanent
-          and can't be undone.
-        </Typography>
-      ) : (
-        <Typography>
-          Deleting a bucket is permanent and can't be undone.
-        </Typography>
-      )}
-      <Typography className={classes.confirmationCopy}>
-        To confirm deletion, type the name of the bucket ({bucketToRemove.label}
-        ) in the field below:
-      </Typography>
-    </React.Fragment>
-  ) : null;
+  const {
+    data,
+    orderBy,
+    order,
+    handleOrderChange,
+    handleClickRemove,
+    classes
+  } = props;
 
   return (
     <Paginate data={data} pageSize={25}>
@@ -227,27 +125,6 @@ export const ListBuckets: React.StatelessComponent<CombinedProps> = props => {
             handleSizeChange={handlePageSizeChange}
             eventCategory="object storage landing"
           />
-          <ConfirmationDialog
-            open={removeBucketConfirmationDialog.isOpen}
-            onClose={() => {
-              removeBucketConfirmationDialog.close();
-            }}
-            title={
-              bucketToRemove
-                ? `Delete ${bucketToRemove.label}`
-                : 'Delete bucket'
-            }
-            actions={actions}
-            error={error}
-          >
-            {deleteBucketConfirmationMessage}
-            <TextField
-              onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-                setConfirmBucketName(e.target.value)
-              }
-              expand
-            />
-          </ConfirmationDialog>
         </React.Fragment>
       )}
     </Paginate>
@@ -277,9 +154,6 @@ const RenderData: React.StatelessComponent<RenderDataProps> = props => {
 
 const styled = withStyles(styles);
 
-const enhanced = compose<CombinedProps, Props>(
-  styled,
-  bucketRequestsContainer
-);
+const enhanced = compose<CombinedProps, Props>(styled);
 
 export default enhanced(ListBuckets);
